Fix requiring generic parameters to implement Default
Currently the builder uses #[derive(Default)], which adds additional bounds to the type parameters
requiring them to implement Default. This is not required since the fields are all either `Option`
or `PhantomData`, which always have defaults, and means that Builders cannot be instantiated for
non-Default generics.

This commit makes the builder implement Default manually, whithout any additional bounds.