Remove skeptic
Complexity in the repo's tests have made it difficult for maintainers
and contributors to work on the crate with confidence that they haven't
caused regressions. `skeptic` was a major driver of that complexity, as
it introduced the need for an additional piece of the test matrix and
custom build scripts.

That complexity was only testing a single example in the README, and
that example is already covered by the examples/ directory. Removing
skeptic simplifies the repo and the CI test pass with minimal reduction
in test coverage.