Shrink derive_builder_core's public API
The only crate to depend on derive_builder_core is derive_builder.
While it's useful that derive_builder_core is _able_ to export more than
the derive function in the future, shrinking the API surface to be only
the one function now means that adding builder features won't force
semver bumps due to changes in the public-facing intermediate structs.

Because the structs are no longer public, doc-tests for those structs
no longer compile. These are now ignored.