Attach validation-related errors to better span
If the build() method cannot convert the validation error to the
generated builder error, this is a problem with the validation function.
The error span should point there, not at the macro call-site.

Due to some issue in `darling` or `syn`, this change is pointing at the
closing quotation instead of the whole path; that is being tracked as
TedDriggs/darling#120.