Safely support custom errors (#194)
* Add UninitializedFieldError

Structured errors need a safe way to act on uninitialized fields
that doesn't conflate with validation errors. This commit adds that by
providing a struct to represent an uninitialized field and generating
conversions from that into the auto-generated error type.

* Attach validation-related errors to better span

If the build() method cannot convert the validation error to the
generated builder error, this is a problem with the validation function.
The error span should point there, not at the macro call-site.

* Allow specification of preexisting error

Using `#[builder(build_fn(error = "..."))]` allows passing an error type 
to use instead of generating one. This makes it possible to integrate with
a crate's existing error.

* Add run-pass custom error test

This test checks that we don't require unnecessary impls
for custom errors when uninitialized fields are impossible.

Fixes #181 
Fixes #191