Only generate ValidationError when needed
Custom validation errors are converted to strings if the deriving struct
doesn't specify a type for the errors returned by the `build` method.
This variant was generated even when it wasn't needed, resulting in a
reference to `String` that broke no_std.

This change only emits that variant when `build_fn(validate = "...")` is
set **and** `build_fn(error = "...")` is not set. This means no_std will
work in the simple case with no additional meta items needed.

Fixes #201