Support setter(strip_option, each = "name")
Currently, combining setter(strip_option) and setter(each) to provide
an "append" API for an optional field fails to compile due to the
setter(each) implementation using the original field type in the trait
bounds for Extend. When setter(strip_option) is used, the original field
will be Option<collection>. This commit fixes this by "traversing" the
two levels of Option (initialising with defaults as necessary) to change
the trait bounds for Extend to the inner collection type.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>