Introduce BuilderFieldType enum
We are going to make some fields in the builder be a bare value.  This
enum allows us to avoid accidentally using the field the wrong way.
It impl ToTokens, which adds the Option (where applicable), and
provides access to the target field type (i.e., usually, the version
without the Option).

Right now there is only one variant - no functional change yet.