Attribute pass-through for builder struct and impl block
Following on from the addition of attribute pass-through for fields
and setters, it's now possible to specify attributes for the struct
using `#[builder_struct_attr(...)]` and for the inherent impl block
using `#[builder_impl_attr(...)]`.

With this change, container-level `serde` attributes are supported
and `#[wasm_bindgen]` is possible, fixing #221 (though making the
builder work with WASM remains the caller's responsibility).

Fixes #239