Abolish BuilderFieldType::target_type in favour of setter_type_info
target_type was wrong; this is not necessarily the target field type:
for PhantomData it isn't.  At the one call site, this didn't matter,
because it could never happen.

wrap_expression_in_some was rather funky.

Instead, use a pattern similar to strip_option, where we record
in a bool whether to nest in an option.