Produce error for `default` with `field.type`
If `field.type` is used without `field.build`, a Move conversion is used instead of the OptionOrDefault conversion.
This means that the presence of `field.type` prevents a field-level `default` from doing anything.
Standard practice in derive_builder is to produce an error when parameters are not going to be used.

Fixes #269