Use darling's `with` to handle unnesting attrs
Before darling 0.20.8, the receiver struct had to take the forwarded attributes without transformation.
That led to the options structs having a field called attrs that was drained during FromDeriveInput/FromField execution, and was vestigial on the finished struct.

The new approach instead does that transformation in the attrs field, resulting in an emitted type that is cleaner and more obvious to use.